Refactor API controller and move code to Account controller/service
