+++ apps/api/src/controllers/accounts.controller.ts
@@ -1,7 +1,9 @@
 import { Body, Controller, Get, Post, HttpCode, HttpStatus, Logger, Param, HttpException } from '@nestjs/common';
 import { ApiBody, ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { AccountsService } from '../services/accounts.service';
-import { Account, AccountResponse } from '../../../../libs/common/src/types/dtos/accounts.dto';
+import { AccountResponse } from '../../../../libs/common/src/types/dtos/accounts.dto';
+import { WalletLoginResponseDTO } from '../../../../libs/common/src/types/dtos/wallet.login.response.dto';
+import { WalletLoginRequestDTO } from '../../../../libs/common/src/types/dtos/wallet.login.request.dto';
 
 @Controller('accounts')
 @ApiTags('accounts')
@@ -30,6 +32,26 @@ export class AccountsController {
     } catch (error) {
       this.logger.error(error);
       throw new HttpException('Failed to find the account', HttpStatus.BAD_REQUEST);
+    }
+  }
+
+  @Post('login')
+  @HttpCode(HttpStatus.CREATED)
+  @ApiOperation({ summary: 'Request to sign in with Frequency' })
+  @ApiOkResponse({ description: 'Signed in successfully', type: WalletLoginResponseDTO })
+  @ApiBody({ type: WalletLoginRequestDTO })
+  async signInWithFrequency(
+    @Body() walletLoginRequestDTO: WalletLoginRequestDTO,
+  ): Promise<WalletLoginResponseDTO | Response> {
+    try {
+      this.logger.log('Received Sign-In With Frequency request');
+      this.logger.debug(`walletLoginRequestDTO: ${JSON.stringify(walletLoginRequestDTO)}`);
+      const loginResponse = await this.accountsService.signInWithFrequency(walletLoginRequestDTO);
+      // return { status: 202, message: `SIWF in progress. referenceId: ${referenceId}`, data: loginResponse };
+      return loginResponse;
+    } catch (error) {
+      this.logger.error(`Failed to Sign In With Frequency: ${error}`);
+      throw new HttpException('Failed to Sign In With Frequency', HttpStatus.BAD_REQUEST);
     }
   }
 }
